requirements.txt
flask==3.0.3
flask-login==0.6.3
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.32
python-dotenv==1.0.1
markdown2==2.5.0
bleach==6.1.0
gunicorn==22.0.0
reportlab==4.2.0
